refactor(ipc): extract auth-oauth-handlers from sync-handlers (Phase 2.1)
Split OAuth loopback, PKCE state, and OAuth-flow handlers out of the
1157-line sync-handlers god module into a cohesive auth-oauth-handlers
module with matching tests. sync-handlers continues to compose the new
module's register/unregister into its public API so ipc/index.ts and
session-teardown consumers stay unchanged this commit.

Moved: AUTH_INIT_OAUTH, AUTH_REFRESH_TOKEN, SETUP_FIRST_DEVICE,
CONFIRM_RECOVERY_PHRASE, GET_RECOVERY_PHRASE, AUTH_LOGOUT + oauthSessions
map, loopback server helpers, and seedOAuthSession export. GET_RECOVERY_PHRASE
consumes pendingRecoveryPhrase via a narrow getter from auth-device-handlers
(still in sync-handlers for now; will land in Phase 2.2).

- sync-handlers.ts 1157 → 866 lines
- sync-handlers.test.ts 749 → 554 lines
- New: auth-oauth-handlers.ts (364), auth-oauth-handlers.test.ts (341)
